main/FirstFlaskWebApp/requirements.txt
Flask==2.1.3
h5py==3.1.0
imgaug==0.4.0
ipython==8.6.0
keras==2.11.0
keras_nightly==2.5.0.dev2021032900
matplotlib==3.6.0
numpy==1.19.5
opencv_python==4.6.0.66
pandas==1.3.3
pdf2image==1.16.0
Pillow==9.3.0
pytesseract==0.3.10
scikit_image==0.16.2
scipy==1.9.1
skimage==0.0
tabulate==0.9.0
tensorflow==2.11.0
tensorflow_gpu==2.5.0
